Start of Game UI Implementation. Trials Implementation.
The player is now able to select the amount of trials they would like to
play, ranging from 1-9.
There are UI prompts for this selection process.
The player can now also play through multiple rounds. At the moment, it
is confusing as to when a new round has actually started.
An indication of such will be necessary.
Also, if the player gets to the end of the trial and hasn't pressed
space, there will need to be a 'loss' prompt.