Big Update
Scoring has been refined and is more accurate.
There is a winner state, and an end-game state.
The winner state follows the end of a singular trial, while the end-game
proceeds the final trial.
The end-game state provides the player with an average of all scores
throughout the trials played.
Currently there is a bug that spawned that involves transitions to
proceeding elements in the list.