Finished
Fixed major bug that was stopping the game from playing properly. Some
logic was off. The coroutine was running more than it should have been.
Game now has messages that respond to an incorrect press as well as if
the amount of icons runs out.
Trials will finish properly
Transitions happen in a uniform fashion, always.